# backend/openapi.yaml
swagger: "2.0"
info:
  title: E-commerce API
  description: API for E-commerce platform
  version: "1.0.0"
host: "intersport-gateway-2abu0fa1.uc.gateway.dev"
...
paths:
...
  /addToCart:
    post:
      summary: Add item to cart
      operationId: addToCart
      x-google-backend:
        address: https://us-central1-gcp-atlas.cloudfunctions.net/addToCart
      parameters:
        - in: body
          name: body
          required: true
          schema:
            type: object
            properties:
              userId:
                type: string
              productId:
                type: string
              color:
                type: string
              size:
                type: string
              quantity:
                type: integer
      responses:
        "200":
          description: Item added to cart
    options:
      summary: CORS support for addToCart
      operationId: corsAddToCart
      x-google-backend:
        address: https://us-central1-gcp-atlas.cloudfunctions.net/addToCart
      responses:
        "200":
          description: CORS headers sent
...